requirements.txt
# Core Data Processing
pandas>=1.5.0
numpy>=1.24.0
scikit-learn>=1.2.0

# API Development  
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0

# Data Download
requests>=2.31.0

# Jupyter Notebooks
jupyter>=1.0.0

# Visualization (optional)
matplotlib>=3.6.0
seaborn>=0.12.0

# Database (built into Python)
# sqlite3 - no need to install

mlflow>=2.7.0
dvc>=3.0.0 
prometheus-client>=0.17.0
evidently>=0.4.0

# MySQL
pymysql        
psycopg2-binary
mysql-connector-python
python-dotenv
SQLAlchemy>=2.0.0